+++ docs/changelog.rst
@@ -3,5 +3,5 @@ Changelog
 
 * **v0.0 ()**
 
-  * Added ``Game``, ``Scene``, ``Object``, ``Screen``, ``Sprite``
+  * Added ``Game``, ``Scene``, ``Object``, ``Screen``, ``Sprite``, ``Text``
   * Added ``tegen.pixel``